second review3

+++ backend/api/mixins.py
@@ -1,7 +1,8 @@
 from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
 from rest_framework.serializers import ModelSerializer
-from rest_framework.status import HTTP_201_CREATED, HTTP_204_NO_CONTENT
+from rest_framework.status import (
+    HTTP_201_CREATED, HTTP_204_NO_CONTENT, HTTP_400_BAD_REQUEST)
 from rest_framework.viewsets import ModelViewSet
 
 from recipes.models import IngredientInRecipe, Recipe
@@ -23,9 +24,12 @@ class CreateDestroy(ModelViewSet):
     def delete_method_for_actions(request, pk, model):
         user = request.user
         recipe = get_object_or_404(Recipe, id=pk)
-        model_obj = get_object_or_404(model, user=user, recipe=recipe)
-        model_obj.delete()
-        return Response(status=HTTP_204_NO_CONTENT)
+        model_obj = model.objects.filter(user=user, recipe=recipe)
+        if model_obj.exists():
+            model_obj.delete()
+            return Response(status=HTTP_204_NO_CONTENT)
+        return Response(
+            {'error': 'Этого рецепта нет у вас'}, status=HTTP_400_BAD_REQUEST)
 
 
 class IsSubscribed():
